Reorganized actionhud modules
