+++ game/view/gameplay/actionhud/init.lua
@@ -11,7 +11,7 @@ local HandView      = require 'view.gameplay.actionhud.hand'
 local Minimap       = require 'view.gameplay.actionhud.minimap'
 local EquipmentDock = require 'view.gameplay.actionhud.equipmentdock'
 local ConditionDock = require 'view.gameplay.actionhud.conditiondock'
-local FocusBar      = require 'view.focusbar'
+local FocusBar      = require 'view.gameplay.actionhud.focusbar'
 local HoldBar       = require 'view.helpers.holdbar'
 local vec2          = require 'cpml' .vec2
 local Util          = require "steaming.util"
